requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
SQLAlchemy==2.0.28
asyncpg==0.29.0
redis==5.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pytz==2024.1
celery==5.3.6
httpx==0.27.0
aiosmtplib==3.0.1
python-dotenv==1.0.1
websockets==12.0